Regression: If Asterisk suddenly goes down, server has no way to know. Causes server to get stuck. Needs restart (#24624)
* Clickup Task : https://app.clickup.com/t/22qmrw3
Description:
This PR implements a connection retries in case the asterisk connection dies for some reason.
It does 5 retries with backoff options. The retries will not be done for connectivity check.

Once the retries are exhausted, and server comes up, there is no way to connect back but to restart
the voip service. This works only for management connection. Call Server connection, if dies, there are no
retries.

* Grammar rename on vars

Co-authored-by: Kevin Aleman <kevin.aleman@rocket.chat>